pyproject.toml
[tool.poetry]
name = "test_awersome_django"
version = "0.1.0"
description = ""
authors = ["Дороватовский Евгений <eanegin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
django-easy-audit = "^1.3.3"
python-decouple = "^3.6"
dj-database-url = "^1.0.0"
django-tables2 = "^2.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
